PDF2PNG/requirements.txt
Pillow==9.0.1
PyMuPDF==1.18.19
```